requirements.txt
coremltools
Pillow
